requirements.txt
astroid==3.0.2
config==0.5.1
dill==0.3.7
isort==5.13.2
mccabe==0.7.0
platformdirs==4.1.0
psycopg2-binary==2.9.9
pylint==3.0.3
PyYAML==6.0.1
SQLAlchemy==2.0.25
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
